Optional args to support Material Design Widgets
Added to optional args to the functions extractBinding and formatBinding.
I need this to support the Material Design Widget Theme Editor comming with v0.4.0.

At the moment i just copied these functions to the adapter it self. But i think for long term support it would be better to have this optional args directly in the original functions.